[ALPS04262735] ext4: always verify the magic number in xattr blocks
If there an inode points to a block which is also some other type of
metadata block (such as a block allocation bitmap), the
buffer_verified flag can be set when it was validated as that other
metadata block type; however, it would make a really terrible external
attribute block.  The reason why we use the verified flag is to avoid
constantly reverifying the block.  However, it doesnt take much
overhead to make sure the magic number of the xattr block is correct,
and this will avoid potential crashes.

This addresses CVE-2018-10879.

https://bugzilla.kernel.org/show_bug.cgi?id=200001

MTK-Commit-Id: 90fd028ec3b8afc2f977b4c375a92afe72302267

Change-Id: If1998f249554766216db836bba99e50e6bbd773a
Signed-off-by: Theodore Tso <tytso@mit.edu>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Light Hsieh <light.hsieh@mediatek.com>
CR-Id: ALPS04262735
Feature: [Android Default] EXT4 File System